refactor(proposals): split api/proposals.py into focused package
